Add error checking (Request from Zabuzard)

+++ application/src/main/java/org/togetherjava/tjbot/commands/bytecode/BytecodeCommand.java
@@ -119,6 +119,10 @@ public final class BytecodeCommand implements EventListener {
     }
 
     private void deleteMyMessages(@NotNull Long msgId, @NotNull TextChannel channel) {
+        if (!userMessageToMyMessages.containsKey(msgId)) {
+            return;
+        }
+
         userMessageToMyMessages.get(msgId)
             .forEach(id -> channel.retrieveMessageById(id).queue(msg -> msg.delete().queue()));
